pyproject.toml
[project]
name = "nudget_ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "duckduckgo-search>=8.0.1",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-google-genai>=2.1.3",
    "langchain-huggingface>=0.1.2",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.3.34",
    "playwright>=1.51.0",
    "pydantic[email]>=2.11.3",
    "python-dotenv>=1.1.0",
    "requests-cache>=1.2.1",
    "streamlit>=1.44.1",
]
